✨Release 1.0 `interface`

+++ src/components/UploadFile/index.tsx
@@ -1,34 +1,45 @@
-import { toast } from "react-toastify";
-import { useCallback, useEffect, useMemo } from "react";
-import { useDispatch, useSelector } from "react-redux";
-
-import { Box, Button, LinearProgress, Typography } from "@mui/material";
-import UploadIcon from "@mui/icons-material/Upload";
 import DownloadIcon from "@mui/icons-material/Download";
-
-import { useDropzone } from "react-dropzone";
+import UploadIcon from "@mui/icons-material/Upload";
+import { Box, Button, LinearProgress, Typography } from "@mui/material";
 import { BeforeUnload } from "components";
-
+import {
+  COMPRESS_FILE_EXTENSIONS,
+  ID_TOKEN_NAME,
+  IMAGE_EXTENSIONS,
+  LIMIT_IMAGE_HEIGHT,
+  LIMIT_IMAGE_WIDTH,
+  MAX_ALLOW_UPLOAD_IMAGES,
+  MAX_ALLOW_UPLOAD_IMAGES_AT_THE_SAME_TIME,
+} from "constants/defaultValues";
+import {
+  ADDED_UPLOAD_FILE_STATUS,
+  INVALID_FILE_STATUS,
+  UPLOADED_UPLOAD_FILE_STATUS,
+} from "constants/uploadFile";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useDropzone } from "react-dropzone";
+import { useDispatch, useSelector } from "react-redux";
+import { toast } from "react-toastify";
+import { selectorIsAlbumSelectMode } from "reduxes/album/selector";
+import {
+  selectorIsGenerateImagesAugmenting,
+  selectorIsGenerateImagesPreprocessing,
+} from "reduxes/generate/selector";
+import {
+  selectorCurrentProjectTotalOriginalImage,
+  selectorHaveTaskRunning,
+} from "reduxes/project/selector";
 import {
   addFile,
   checkFileUpload,
   clearAllFile,
+  clearFileArray,
   deleteFile,
   resetUploadState,
   setTotalUploadFileQuantity,
+  updateFile,
   uploadFile,
 } from "reduxes/upload/actions";
-
-import { getLocalStorage } from "utils/general";
-import {
-  ID_TOKEN_NAME,
-  MAX_ALLOW_UPLOAD_IMAGES,
-} from "constants/defaultValues";
-
-import {
-  ADDED_UPLOAD_FILE_STATUS,
-  UPLOADED_UPLOAD_FILE_STATUS,
-} from "constants/uploadFile";
 import {
   selectorAddedStatusFileCount,
   selectorIsChecking,
@@ -39,19 +50,11 @@ import {
   selectorUploadFiles,
   selectorUploadingFileCount,
 } from "reduxes/upload/selector";
-import {
-  selectorCurrentProjectTotalOriginalImage,
-  selectorHaveTaskRunning,
-} from "reduxes/project/selector";
-import { selectorIsAlbumSelectMode } from "reduxes/album/selector";
-import { UploadFileProps } from "./type";
+import { getLocalStorage, isImageFile } from "utils/general";
+import { LoadImageResult, MousePosition, UploadFileProps } from "./type";
 import UploadFileItem from "./UploadFileItem";
-
 import UploadFromMenu from "./UploadFromMenu";
-import {
-  selectorIsGenerateImagesAugmenting,
-  selectorIsGenerateImagesPreprocessing,
-} from "reduxes/generate/selector";
+import UploadGuideDialog from "./UploadGuideDialog";
 
 interface HandleFileType {
   name: string;
@@ -85,7 +88,10 @@ const dropZoneBaseStyle = {
 const dropZoneDisabledStyle = {
   opacity: 0.5,
 };
-
+const FILE_UPLOAD_EXTENSIONS = [
+  ...IMAGE_EXTENSIONS,
+  ...COMPRESS_FILE_EXTENSIONS,
+];
 const UploadFile = function (props: UploadFileProps) {
   const { projectId, projectName } = props;
 
@@ -124,7 +130,7 @@ const UploadFile = function (props: UploadFileProps) {
   const totalOriginalImage = useSelector(
     selectorCurrentProjectTotalOriginalImage
   );
-
+  const [isOpenUploadGuideDialog, setIsOpenUploadGuideDialog] = useState(false);
   const isDisabledUpload = useMemo(
     () =>
       isUploadChecking ||
@@ -151,9 +157,31 @@ const UploadFile = function (props: UploadFileProps) {
         ).toFixed(0)
       : 0
   );
-
+  const loadImage = (file: File, fileName: string) =>
+    new Promise<LoadImageResult>((resolve) => {
+      const image = new Image();
+      const objectUrl = window.URL.createObjectURL(file);
+      image.onload = () => {
+        resolve({
+          image,
+          fileName,
+        });
+        window.URL.revokeObjectURL(objectUrl);
+      };
+      image.src = objectUrl;
+    });
   const onDrop = useCallback((acceptedFiles) => {
     if (acceptedFiles && acceptedFiles.length > 0) {
+      if (acceptedFiles.length > MAX_ALLOW_UPLOAD_IMAGES_AT_THE_SAME_TIME) {
+        toast.error(
+          <p>
+            The maximum number of uploaded images in a project at the same time
+            is {MAX_ALLOW_UPLOAD_IMAGES_AT_THE_SAME_TIME}. Please, re-upload
+            with less than the number of image images.
+          </p>
+        );
+        return;
+      }
       if (acceptedFiles.length + totalOriginalImage > MAX_ALLOW_UPLOAD_IMAGES) {
         toast.warning(
           <p>
@@ -163,38 +191,64 @@ const UploadFile = function (props: UploadFileProps) {
           </p>
         );
       }
-      dispatch(addFile({ files: convertArrayToObjectKeyFile(acceptedFiles) }));
+      const formatedFiles = convertArrayToObjectKeyFile(acceptedFiles);
+      const files = { ...uploadFiles, ...formatedFiles };
+      dispatch(addFile({ files: formatedFiles }));
+
+      if (files && Object.keys(files).length > 0) {
+        const filesNeedCheck: Array<string> = [];
+        const listPromiseLoadImages: Promise<LoadImageResult>[] = [];
+        Object.keys(files).forEach((fileName: string) => {
+          if (files[fileName].status !== UPLOADED_UPLOAD_FILE_STATUS) {
+            filesNeedCheck.push(fileName);
+            if (isImageFile(fileName)) {
+              listPromiseLoadImages.push(
+                loadImage(files[fileName].file, fileName)
+              );
+            }
+          }
+        });
+        Promise.all(listPromiseLoadImages).then((values) => {
+          values.forEach(({ image, fileName }) => {
+            if (
+              image.width > LIMIT_IMAGE_WIDTH ||
+              image.height > LIMIT_IMAGE_HEIGHT
+            ) {
+              dispatch(
+                updateFile({
+                  fileName,
+                  updateInfo: {
+                    error: `The image resolution (${image.width} x ${image.height}) exceeds the limit allowed ${LIMIT_IMAGE_WIDTH} x ${LIMIT_IMAGE_HEIGHT}. Please remove it`,
+                    status: INVALID_FILE_STATUS,
+                  },
+                })
+              );
+
+              const indexOf = filesNeedCheck.indexOf(fileName);
+              if (indexOf !== -1) {
+                filesNeedCheck.splice(indexOf, 1);
+              }
+            }
+          });
+          dispatch(
+            checkFileUpload({
+              idToken: getLocalStorage(ID_TOKEN_NAME) || "",
+              projectId,
+              projectName,
+              listFileName: filesNeedCheck,
+            })
+          );
+        });
+      }
     }
   }, []);
-
-  const onClickUpload = async () => {
-    if (uploadFiles && Object.keys(uploadFiles).length > 0) {
-      const filesNeedCheck: Array<string> = [];
-
-      Object.keys(uploadFiles).forEach((fileName: string) => {
-        if (uploadFiles[fileName].status !== UPLOADED_UPLOAD_FILE_STATUS) {
-          filesNeedCheck.push(fileName);
-        }
-      });
-
-      dispatch(
-        checkFileUpload({
-          idToken: getLocalStorage(ID_TOKEN_NAME) || "",
-          projectId,
-          projectName,
-          listFileName: filesNeedCheck,
-        })
-      );
-    }
-  };
-
   const onClickClearAll = () => {
     dispatch(clearAllFile());
   };
 
   const { getRootProps, getInputProps, isDragActive, inputRef } = useDropzone({
     onDrop,
-    accept: "image/png, image/jpeg",
+    accept: FILE_UPLOAD_EXTENSIONS.join(","),
     disabled: isDisabledUpload,
   });
 
@@ -208,7 +262,20 @@ const UploadFile = function (props: UploadFileProps) {
 
   const onDeleteFile = (fileName: string) => {
     if (uploadFiles && uploadFiles[fileName]) {
-      dispatch(deleteFile({ fileName }));
+      let isUploadSucess = true;
+      Object.keys(uploadFiles).forEach((name) => {
+        if (
+          name !== fileName &&
+          uploadFiles[name].status !== UPLOADED_UPLOAD_FILE_STATUS
+        )
+          isUploadSucess = false;
+      });
+      if (isUploadSucess) {
+        dispatch(clearFileArray({ fileNameArray: Object.keys(uploadFiles) }));
+        dispatch(setTotalUploadFileQuantity({ totalUploadFileQuantity: null }));
+      } else {
+        dispatch(deleteFile({ fileName }));
+      }
     }
   };
 
@@ -221,6 +288,7 @@ const UploadFile = function (props: UploadFileProps) {
         projectName,
         isReplace: true,
         isReplaceSingle: true,
+        isExist: true,
       })
     );
   };
@@ -231,6 +299,22 @@ const UploadFile = function (props: UploadFileProps) {
     },
     []
   );
+  const [isOpenChooseFileMenu, setIsOpenChooseFileMenu] =
+    useState<boolean>(false);
+  const [relativeMousePosition, setRelativeMousePosition] =
+    useState<MousePosition>({ top: 0, left: 0 });
+  const handleCloseChooseFileMenu = () => {
+    setIsOpenChooseFileMenu(false);
+  };
+  const dropUploadAreaRef = useRef<HTMLElement | null>();
+
+  const handleClickDropUploadArea = (event: React.MouseEvent<HTMLElement>) => {
+    const rect = event.currentTarget.getBoundingClientRect();
+    const x = event.clientX - rect.left;
+    const y = event.clientY - rect.top;
+    setRelativeMousePosition({ top: y, left: x });
+    setIsOpenChooseFileMenu(true);
+  };
 
   const renderDropzoneContent = () => (
     <Box
@@ -249,13 +333,17 @@ const UploadFile = function (props: UploadFileProps) {
           ? "Drop your images here!"
           : "Drop your images here or click to upload"}
       </Typography>
-      <Typography fontStyle="italic" variant="body2" color="text.secondary">
-        .png, .jpg, .jpeg
+      <Typography
+        fontStyle="italic"
+        variant="body2"
+        color="text.secondary"
+        textAlign="center"
+      >
+        Supported file formats: {FILE_UPLOAD_EXTENSIONS.join(", ")}
       </Typography>
       <Typography
         sx={{ mt: 1 }}
         textAlign="center"
-        fontStyle="italic"
         variant="body2"
         color="text.secondary"
       >
@@ -288,6 +376,12 @@ const UploadFile = function (props: UploadFileProps) {
     return null;
   };
 
+  const handleClickShowUploadGuideDialog = () => {
+    setIsOpenUploadGuideDialog(true);
+  };
+  const handleCloseUploadGuideDialog = () => {
+    setIsOpenUploadGuideDialog(false);
+  };
   return (
     <Box my={4}>
       <BeforeUnload
@@ -304,13 +398,18 @@ const UploadFile = function (props: UploadFileProps) {
       )}
 
       <Box display="flex" justifyContent="center" alignItems="stretch">
-        <Box flex={1}>
-          <Box>
-            <UploadFromMenu
-              isDisabledUpload={isDisabledUpload}
-              inputRef={inputRef}
-            />
-          </Box>
+        <Box>
+          <UploadFromMenu
+            isDisabledUpload={isDisabledUpload}
+            inputRef={inputRef}
+            anchorEl={dropUploadAreaRef?.current}
+            relativeMousePosition={relativeMousePosition}
+            isOpen={isOpenChooseFileMenu}
+            onClose={handleCloseChooseFileMenu}
+          />
+        </Box>
+
+        <Box flex={1} ref={dropUploadAreaRef}>
           <Box
             sx={{
               width: "100%",
@@ -326,12 +425,34 @@ const UploadFile = function (props: UploadFileProps) {
             display="flex"
             alignItems="center"
             justifyContent="center"
-            {...getRootProps({ style: dropZoneStyle })}
+            {...getRootProps({
+              onClick: (event) => {
+                handleClickDropUploadArea(event);
+                event.stopPropagation();
+              },
+              style: dropZoneStyle,
+            })}
             mr={2}
           >
             <input {...getInputProps()} multiple />
             {renderDropzoneContent()}
           </Box>
+          {/* <Typography mt={1} fontStyle="italic" variant="body2" fontSize={14}>
+            * If you would like to upload your dataset programmatically,{" "}
+            <Typography
+              component="span"
+              sx={{ cursor: "pointer", fontWeight: "bold", color: "#1c68dc" }}
+              onClick={handleClickShowUploadGuideDialog}
+              fontSize={14}
+            >
+              click here{" "}
+            </Typography>
+            for more information.
+          </Typography> */}
+          <UploadGuideDialog
+            onClose={handleCloseUploadGuideDialog}
+            isOpen={isOpenUploadGuideDialog}
+          />
         </Box>
 
         <Box flex={2} ml={2}>
@@ -352,17 +473,6 @@ const UploadFile = function (props: UploadFileProps) {
               onClick={onClickClearAll}
             >
               Clear All
-            </Button>
-            <Button
-              sx={{ ml: "auto" }}
-              variant="contained"
-              color="primary"
-              onClick={onClickUpload}
-              disabled={
-                uploadFilesLength <= 0 || isUploading || isUploadChecking
-              }
-            >
-              Upload
             </Button>
           </Box>
